# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Inventory Management System REST API. Every response is wrapped in a `data` envelope; errors use an `error` envelope with a stable machine-readable code.",
        "title": "IMS API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/stock/adjust": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Signed quantity: positive receives, negative issues. Supply client_request_id to make the call idempotent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Adjust stock",
                "parameters": [
                    {
                        "description": "Adjustment",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_stock.AdjustInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.MovementResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/stock/batches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "List batches",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Only batches expiring within N days",
                        "name": "filter[expiring_within_days]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_stock.Batch"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Create a batch",
                "parameters": [
                    {
                        "description": "Batch",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_stock.BatchInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.Batch"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/cycle-counts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle-counts"
                ],
                "summary": "List cycle counts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_stock.CycleCount"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle-counts"
                ],
                "summary": "Create a cycle count",
                "parameters": [
                    {
                        "description": "Cycle count scope",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_stock.CycleCountInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.CycleCount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/cycle-counts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle-counts"
                ],
                "summary": "Get a cycle count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle count id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.CycleCount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/cycle-counts/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle-counts"
                ],
                "summary": "Complete a cycle count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle count id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.CycleCount"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/stock/cycle-counts/{id}/record": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cycle-counts"
                ],
                "summary": "Record counted quantities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle count id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Counted lines",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_stock.RecordCountInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.CycleCount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "List stock levels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by warehouse",
                        "name": "filter[warehouse_id]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by product",
                        "name": "filter[product_id]",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only items at or below their minimum",
                        "name": "filter[below_min]",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Exclude zero-quantity rows",
                        "name": "filter[nonzero]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_stock.ItemView"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/movements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "List stock movements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by product",
                        "name": "filter[product_id]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by movement type",
                        "name": "filter[type]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_stock.Movement"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.PageMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/stock/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Applies each queued movement independently and reports a per-item outcome, so one rejected entry does not discard the batch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Sync offline stock movements",
                "parameters": [
                    {
                        "description": "Queued movements",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_stock.SyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sabin-bhattarai_ims-backend_internal_shared.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_stock.SyncResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_stock.AdjustInput": {
            "type": "object",
            "required": [
                "product_id",
                "reason",
                "warehouse_id"
            ],
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "client_request_id": {
                    "description": "ClientRequestID makes a retried offline scan idempotent.",
                    "type": "string",
                    "maxLength": 64
                },
                "location_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string",
                    "maxLength": 1000
                },
                "occurred_at": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "description": "Quantity is signed: positive adds, negative removes.",
                    "type": "number"
                },
                "reason": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "unit_cost": {
                    "type": "number"
                },
                "variant_id": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token as \"Bearer \u003ctoken\u003e\". Obtain one from /auth/login and refresh it at /auth/refresh.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}